Refactor the `Span` implementation
It's likely to get large.
